requirements.txt
fastapi>=0.68.0,<0.69.0
pydantic>=1.8.0,<2.0.00
uvicorn>=0.15.0,<0.16.0
email-validator>=1.1.3,<1.2.1
databases[mysql]>=0.5.4<0.5.5
passlib[bcrypt]>=1.7.3,<1.7.4
python-dotenv>=0.19.2,<0.20.0
requests
pytest